requirements.txt
yfinance>=0.2.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
flask>=3.0.0
plotly>=5.0.0
pyarrow>=14.0.0
